Refactor nan-sensitive hashmap
