+++ include/onnxruntime/core/common/inlined_containers.h
@@ -61,65 +61,9 @@ class InlinedHashSet : public absl::flat_hash_set<T,
 };
 
 template <typename Key, typename Value,
-          typename Allocator>
-class InlinedHashMap : public absl::flat_hash_map<Key, Value,
-                                                  absl::container_internal::hash_default_hash<Key>,
-                                                  absl::container_internal::hash_default_eq<Key>,
-                                                  Allocator> {
-  using Base = absl::flat_hash_map<Key, Value,
-                                   absl::container_internal::hash_default_hash<Key>,
-                                   absl::container_internal::hash_default_eq<Key>,
-                                   Allocator>;
-
- public:
-  using Base::Base;
-};
-
-template <typename T>
-struct NaNHash {
-  size_t operator()(const T& value) const {
-    return absl::container_internal::hash_default_hash<T>{}(value);
-  }
-};
-
-template <>
-struct NaNHash<float> {
-  size_t operator()(const float value) const {
-    if (std::isnan(value)) {
-      return 0;
-    } else {
-      return absl::container_internal::hash_default_hash<float>{}(value);
-    }
-  }
-};
-
-template <typename T>
-struct NaNEqual {
-  bool operator()(const T& lhs, const T& rhs) const {
-    return absl::container_internal::hash_default_eq<T>{}(lhs, rhs);
-  }
-};
-
-template <>
-struct NaNEqual<float> {
-  bool operator()(const float lhs, const float rhs) const {
-    if (std::isnan(lhs) && std::isnan(rhs)) {
-      return true;
-    }
-    return absl::container_internal::hash_default_eq<float>{}(lhs, rhs);
-  }
-};
-
-template <typename Key, typename Value,
-          typename Allocator>
-class InlinedHashMapNaNSensitive : public absl::flat_hash_map<Key, Value,
-                                                              NaNHash<Key>,
-                                                              NaNEqual<Key>,
-                                                              Allocator> {
-  using Base = absl::flat_hash_map<Key, Value,
-                                   NaNHash<Key>,
-                                   NaNEqual<Key>,
-                                   Allocator>;
+          typename Allocator, typename Hash, typename Equal>
+class InlinedHashMap : public absl::flat_hash_map<Key, Value, Hash, Equal, Allocator> {
+  using Base = absl::flat_hash_map<Key, Value, Hash, Equal, Allocator>;
 
  public:
   using Base::Base;
@@ -174,65 +118,12 @@ class InlinedHashSet : public std::unordered_set<T,
 };
 
 template <typename Key, typename Value,
-          typename Allocator>
+          typename Allocator, typename Hash, typename Equal>
 class InlinedHashMap : public std::unordered_map<Key, Value,
-                                                 std::hash<Key>,
-                                                 std::equal_to<Key>,
+                                                 Hash,
+                                                 Equal,
                                                  Allocator> {
-  using Base = std::unordered_map<Key, Value,
-                                  std::hash<Key>,
-                                  std::equal_to<Key>,
-                                  Allocator>;
-
- public:
-  using Base::Base;
-};
-
-template <typename T>
-struct NaNHash {
-  size_t operator()(const T& value) const {
-    return std::hash<Key>{}(value);
-  }
-};
-
-template <>
-struct NaNHash<float> {
-  size_t operator()(const float value) const {
-    if (std::isnan(value)) {
-      return 0;
-    } else {
-      return std::hash<float>{}(value);
-    }
-  }
-};
-
-template <typename T>
-struct NaNEqual {
-  bool operator()(const T& lhs, const T& rhs) const {
-    return std::equal_to<T>{}(lhs, rhs);
-  }
-};
-
-template <>
-struct NaNEqual<float> {
-  bool operator()(const float lhs, const float rhs) const {
-    if (std::isnan(lhs) && std::isnan(rhs)) {
-      return true;
-    }
-    return std::equal_to<float>{}(lhs, rhs);
-  }
-};
-
-template <typename Key, typename Value,
-          typename Allocator>
-class InlinedHashMapNaNSensitive : public absl::flat_hash_map<Key, Value,
-                                                              NaNHash<Key>,
-                                                              NaNEqual<Key>,
-                                                              Allocator> {
-  using Base = absl::flat_hash_map<Key, Value,
-                                   NaNHash<Key>,
-                                   NaNEqual<Key>,
-                                   Allocator>;
+  using Base = std::unordered_map<Key, Value, Hash, Equal, Allocator>;
 
  public:
   using Base::Base;
@@ -273,4 +164,3 @@ class NodeHashMap : public std::unordered_map<Key, Value,
 #endif  // DISABLE_ABSEIL
 
 }  // namespace onnxruntime
-//
